Move all of the backup key stuff into a module.
This is an intermediate state. Eventually the Hsm type won't touch any
of they backup key creation and splitting. The main module will take
over that part and the Hsm module will continue to deminish.
